Better handling of updates available for qui-updates and Qubes Update
Fixed incorrect count of VMs needing an update and added a bunch of QOL
improvements that make it clearer that what checking a VM to be updated
means.
This may be related to a bunch of errors with updates being incorrectly
reported as unavailable.

references QubesOS/qubes-issues#4667
references QubesOS/qubes-issues#4650