[WIP] redo updater; add menu styling
Instead of starting to update devices upon receiving a change event
and dropping subsequent ones, try to empty the queue first, so the
menu reflects the latest state.

Add labels to the device class menu items.

Refresh the menu if it's being shown as changes happen. It can get separated
from the tray bar right now if devices are removed; need to look into
how to reposition it while it's visible.